requirements.txt
blinker==1.4
colorama==0.4.4
cycler==0.11.0
Flask==2.1.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
fonttools==4.29.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3
zipp==3.7.0
